:hammer: Read scattering data from files
