+++ config/base/scene/simple_desktop.json
@@ -744,23 +744,11 @@
           "model": "Bruneton",
           "modelSettings": {
             "sunAngularRadius": 0.004675,
-            "multiScatteringOrder": 10,
-            "molecules": {
-              "phase": "../share/resources/data/csp-atmospheres/earth_cosmoscout_molecules_phase.csv",
-              "betaSca": "../share/resources/data/csp-atmospheres/earth_cosmoscout_molecules_scattering.csv",
-              "betaAbs": "../share/resources/data/csp-atmospheres/earth_cosmoscout_molecules_absorption.csv",
-              "density": "../share/resources/data/csp-atmospheres/earth_cosmoscout_molecules_density.csv"
-            },
-            "aerosols": {
-              "phase": "../share/resources/data/csp-atmospheres/earth_cosmoscout_aerosols_phase.csv",
-              "betaSca": "../share/resources/data/csp-atmospheres/earth_cosmoscout_aerosols_scattering.csv",
-              "betaAbs": "../share/resources/data/csp-atmospheres/earth_cosmoscout_aerosols_absorption.csv",
-              "density": "../share/resources/data/csp-atmospheres/earth_cosmoscout_aerosols_density.csv"
-            },
-            "ozone": {
-              "betaAbs": "../share/resources/data/csp-atmospheres/earth_cosmoscout_ozone_absorption.csv",
-              "density": "../share/resources/data/csp-atmospheres/earth_cosmoscout_ozone_density.csv"
-            }
+            "phaseTexture": "earth-atmo/phase.tif",
+            "transmittanceTexture": "earth-atmo/transmittance.tif",
+            "irradianceTexture": "earth-atmo/irradiance.tif",
+            "singleScatteringTexture": "earth-atmo/single_aerosols_scattering.tif",
+            "multipleScatteringTexture": "earth-atmo/multiple_scattering.tif"
           }
         },
         "Mars": {
@@ -771,19 +759,11 @@
             "transmittanceTextureWidth": 1024,
             "transmittanceTextureHeight": 512,
             "sunAngularRadius": 0.003054,
-            "multiScatteringOrder": 10,
-            "molecules": {
-              "phase": "../share/resources/data/csp-atmospheres/mars_cosmoscout_molecules_cinematic_phase.csv",
-              "betaSca": "../share/resources/data/csp-atmospheres/mars_cosmoscout_molecules_cinematic_scattering.csv",
-              "betaAbs": "../share/resources/data/csp-atmospheres/mars_cosmoscout_molecules_cinematic_absorption.csv",
-              "density": "../share/resources/data/csp-atmospheres/mars_cosmoscout_molecules_cinematic_density.csv"
-            },
-            "aerosols": {
-              "phase": "../share/resources/data/csp-atmospheres/mars_cosmoscout_aerosols_cinematic_phase.csv",
-              "betaSca": "../share/resources/data/csp-atmospheres/mars_cosmoscout_aerosols_cinematic_scattering.csv",
-              "betaAbs": "../share/resources/data/csp-atmospheres/mars_cosmoscout_aerosols_cinematic_absorption.csv",
-              "density": "../share/resources/data/csp-atmospheres/mars_cosmoscout_aerosols_cinematic_density.csv"
-            }
+            "phaseTexture": "earth-atmo/phase.tif",
+            "transmittanceTexture": "earth-atmo/transmittance.tif",
+            "irradianceTexture": "earth-atmo/irradiance.tif",
+            "singleScatteringTexture": "earth-atmo/single_aerosols_scattering.tif",
+            "multipleScatteringTexture": "earth-atmo/multiple_scattering.tif"
           }
         }
       }
