+++ plugins/csp-atmospheres/atmosphere-preprocessor/CMakeLists.txt
@@ -44,3 +44,5 @@ install(
   TARGETS atmosphere-preprocessor
   RUNTIME DESTINATION "bin"
 )
+
+install(DIRECTORY "shaders" DESTINATION "share/resources")
