+++ plugins/csp-atmospheres/atmosphere-preprocessor/shaders/atmosphere-preprocessing-definitions.glsl
